+++ options/qol.lua
@@ -34,6 +34,11 @@ function TUI:BuildQoLConfig(root, tuiName)
         function() return TUI.db.profile.qol.shortenEnchantStrings end,
         function(_, v) TUI.db.profile.qol.shortenEnchantStrings = v; E:StaticPopup_Show('CONFIG_RL') end)
 
+    qolGen.linkActionBarMouseover = ACH:Toggle(E.NewSign .. "Link AB Mouseover",
+        "When you hover any action bar set to Mouseover, all Mouseover action bars fade in together.", 7, nil, nil, nil,
+        function() return TUI.db.profile.qol.linkActionBarMouseover end,
+        function(_, v) TUI.db.profile.qol.linkActionBarMouseover = v end)
+
     -- Cursor Circle
     root.qol.args.cursorCircle = ACH:Group("Cursor Circle", nil, 2)
     local ccArgs = root.qol.args.cursorCircle.args